+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v2.1.4
+
+* Fixed `substreams-sink-kv serve` not working correctly due to invalid flags.
+
+* Fixed `substreams-sink-kv inject --listen-addr="..."` not listing if sinker finishes the requested block range.
+
 ## v2.1.3
 
 ### Fixed